# server/auth-service/pkg/swagger/api.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Registration Service",
    "version": "1.0.0",
    "contact": {
      "name": "AwesomeXjs",
      "url": "https://github.com/AwesomeXjs",
      "email": "bestawes0mex@gmail.com"
    }
  },
  "tags": [
    {
      "name": "AuthV1"
    }
  ],
  "host": "localhost:8080",
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/api/v1/login": {
      "post": {
        "summary": "Login with email and password",
        "operationId": "AuthV1_Login",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/auth_v1LoginResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/auth_v1LoginRequest"
            }
          }
        ],
        "tags": [
          "AuthV1"
        ]
      }
    },
    "/api/v1/password/{email}": {
      "patch": {
        "operationId": "AuthV1_UpdatePassword",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "type": "object",
              "properties": {}
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "email",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "password": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "tags": [
          "AuthV1"
        ]
      }
    },
    "/api/v1/registration": {
      "post": {
        "summary": "Register new user with email and password, returns access token, refresh token and user id",
        "operationId": "AuthV1_Registration",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/auth_v1RegistrationResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/auth_v1RegistrationRequest"
            }
          }
        ],
        "tags": [
          "AuthV1"
        ]
      }
    },
    "/api/v1/token/{refreshToken}": {
      "get": {
        "summary": "Get access token from refresh token",
        "operationId": "AuthV1_GetAccessToken",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/auth_v1GetAccessTokenResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "refreshToken",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "AuthV1"
        ]
      }
    },
    "/api/v1/validate/{accessToken}": {
      "get": {
        "summary": "Validate access token for authentication",
        "operationId": "AuthV1_ValidateToken",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "type": "object",
              "properties": {}
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "accessToken",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "AuthV1"
        ]
      }
    }
  },
  "definitions": {
    "auth_v1GetAccessTokenResponse": {
      "type": "object",
      "properties": {
        "accessToken": {
          "type": "string"
        },
        "refreshToken": {
          "type": "string"
        }
      }
    },
    "auth_v1LoginRequest": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      }
    },
    "auth_v1LoginResponse": {
      "type": "object",
      "properties": {
        "accessToken": {
          "type": "string"
        },
        "refreshToken": {
          "type": "string"
        },
        "userId": {
          "type": "string"
        }
      }
    },
    "auth_v1RegistrationRequest": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "surname": {
          "type": "string"
        },
        "role": {
          "type": "string"
        }
      }
    },
    "auth_v1RegistrationResponse": {
      "type": "object",
      "properties": {
        "accessToken": {
          "type": "string"
        },
        "refreshToken": {
          "type": "string"
        },
        "userId": {
          "type": "string"
        }
      }
    },
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    }
  }
}